[app] Add basic routing set up to App
